Updated services to use Integer instead of Long (because of model)

+++ src/main/java/com/eEducation/ftn/service/ClassService.java
@@ -15,7 +15,7 @@ public class ClassService {
 	@Autowired
 	ClassRepository classRepository;
 	
-	public Class findOne(Long id) {
+	public Class findOne(Integer id) {
 		return classRepository.findOne(id);
 	}
 
@@ -31,7 +31,7 @@ public class ClassService {
 		return classRepository.save(Class);
 	}
 
-	public void remove(Long id) {
+	public void remove(Integer id) {
 		classRepository.delete(id);
 	}
 }
